requirements.txt
tqdm==4.66.1
plyfile==0.8.1
torchmetrics==0.11.4
imagesize==1.4.1
kornia==0.6.12
omegaconf==2.3.0
dahuffman
imageio
lpips

--extra-index-url=https://pypi.nvidia.com
rmm-cu12==25.2.*
cudf-cu12==25.2.*
cuml-cu12==25.2.*

./simple-knn
./pointops2

git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch